refactor: translate error messages, hide thumbnails for deleted items, update notifier
