+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@graasp/chatbox": "github:graasp/graasp-chatbox.git",
     "@graasp/query-client": "github:graasp/graasp-query-client.git",
+    "@graasp/translations": "github:graasp/graasp-translations.git#1/init",
     "@graasp/ui": "github:graasp/graasp-ui.git",
     "@graasp/utils": "github:graasp/graasp-utils.git",
     "@material-ui/core": "4.12.3",
@@ -23,7 +24,7 @@
     "clsx": "1.1.1",
     "connected-react-router": "6.9.2",
     "http-status-codes": "2.2.0",
-    "i18next": "21.6.5",
+    "i18next": "21.6.12",
     "immutable": "4.0.0",
     "js-cookie": "3.0.1",
     "katex": "0.15.1",
