+++ src/components/table/NameCellRenderer.js
@@ -16,40 +16,46 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const NameCellRenderer = ({ data: item }) => {
-  const classes = useStyles();
-  const { t } = useTranslation();
-  const ThumbnailComponent = Thumbnail({
-    id: item.id,
-    extra: item.extra,
-    maxWidth: 30,
-    maxHeight: 30,
-    alt: t('small thumbnail'),
-    useThumbnail: hooks.useItemThumbnail,
-  });
+const NameCellRenderer = (showThumbnails) => {
+  const Component = ({ data: item }) => {
+    const classes = useStyles();
+    const { t } = useTranslation();
+    const ThumbnailComponent = showThumbnails
+      ? Thumbnail({
+          id: item.id,
+          extra: item.extra,
+          maxWidth: 30,
+          maxHeight: 30,
+          alt: t('small thumbnail'),
+          useThumbnail: hooks.useItemThumbnail,
+        })
+      : null;
 
-  return (
-    <div className={classes.nameCell} id={buildNameCellRendererId(item.id)}>
-      {ThumbnailComponent ?? (
-        <ItemIcon
-          type={item.type}
-          id={item.id}
-          extra={item.extra}
-          name={item.name}
-        />
-      )}
-      <div className={classes.name}>{item.name}</div>
-    </div>
-  );
-};
+    return (
+      <div className={classes.nameCell} id={buildNameCellRendererId(item.id)}>
+        {ThumbnailComponent ?? (
+          <ItemIcon
+            type={item.type}
+            id={item.id}
+            extra={item.extra}
+            name={item.name}
+          />
+        )}
+        <div className={classes.name}>{item.name}</div>
+      </div>
+    );
+  };
 
-NameCellRenderer.propTypes = {
-  data: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    extra: PropTypes.shape({}).isRequired,
-    type: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-  }).isRequired,
+  Component.propTypes = {
+    data: PropTypes.shape({
+      id: PropTypes.string.isRequired,
+      extra: PropTypes.shape({}).isRequired,
+      type: PropTypes.string.isRequired,
+      name: PropTypes.string.isRequired,
+    }).isRequired,
+  };
+
+  return Component;
 };
 
 export default NameCellRenderer;
