+++ src/config/messages.js
@@ -1,22 +1,5 @@
-export const GET_SHARED_ITEMS_ERROR_MESSAGE =
-  'There was an error fetching items.';
-export const GET_ITEM_ERROR_MESSAGE = 'There was an error fetching the item.';
-export const CREATE_ITEM_ERROR_MESSAGE =
-  'There was an error creating the item.';
-export const DELETE_ITEMS_ERROR_MESSAGE =
-  'There was an error deleting the item(s).';
-export const MOVE_ITEM_ERROR_MESSAGE = 'There was an error moving the item.';
-export const MOVE_ITEMS_ERROR_MESSAGE = 'There was an error moving the items.';
-export const COPY_ITEM_ERROR_MESSAGE = 'There was an error copying the item.';
-export const COPY_ITEMS_ERROR_MESSAGE = 'There was an error copying the items.';
-export const EDIT_ITEM_ERROR_MESSAGE = 'There was an error editing the item.';
-export const RESTORE_ITEMS_ERROR_MESSAGE =
-  'There was an error restoring the item(s).';
 export const RESTORE_ITEMS_SUCCESS_MESSAGE =
   'You successfully restore the item(s).';
-export const SHARE_ITEM_ERROR_MESSAGE = 'There was an error sharing the item.';
-export const UPLOAD_FILES_ERROR_MESSAGE =
-  'There was an error uploading the file(s).';
 export const CREATE_ITEM_SUCCESS_MESSAGE = 'The item was successfully created.';
 export const DELETE_ITEMS_SUCCESS_MESSAGE =
   'The item(s) were successfully deleted.';
@@ -34,44 +17,22 @@ export const UPLOAD_FILES_SUCCESS_MESSAGE =
   'The file(s) were successfully uploaded.';
 export const UPLOAD_FILES_PROGRESS_MESSAGE =
   'The file(s) are in queue for uploading. Please wait a moment.';
-export const SIGN_OUT_ERROR_MESSAGE = 'There was an error while signing out.';
 export const SIGN_OUT_SUCCESS_MESSAGE = 'You successfully signed out.';
 
-// todo: customize settings
-export const DELETE_ITEM_TAG_ERROR_MESSAGE =
-  'There was an error while deleting the tag.';
-export const POST_ITEM_TAG_ERROR_MESSAGE =
-  'There was an error while posting the tag.';
-export const ITEM_LOGIN_SIGN_IN_ERROR_MESSAGE =
-  'There was an error while signing in.';
-export const EDIT_MEMBER_ERROR_MESSAGE =
-  'There was an error updating the member';
 export const EDIT_MEMBER_SUCCESS_MESSAGE =
   'The member was updated successfully';
 export const COPY_MEMBER_ID_TO_CLIPBOARD_SUCCESS_MESSAGE =
   'Member ID is successfully copied!';
-export const COPY_MEMBER_ID_TO_CLIPBOARD_ERROR_MESSAGE =
-  'An error occured while copying the member ID';
-export const EDIT_ITEM_MEMBERSHIP_ERROR_MESSAGE =
-  'There was an error editing the item membership.';
-export const DELETE_ITEM_MEMBERSHIP_ERROR_MESSAGE =
-  'There was an error deleting the item membership.';
 export const EDIT_ITEM_MEMBERSHIP_SUCCESS_MESSAGE =
   'The item membership was successfully edited.';
 export const DELETE_ITEM_MEMBERSHIP_SUCCESS_MESSAGE =
   'The item membership was successfully deleted.';
 export const POST_ITEM_FLAG_SUCCESS_MESSAGE =
   'The item was successfully flagged';
-export const POST_ITEM_FLAG_ERROR_MESSAGE =
-  'There was an error flagging the item';
 export const COPY_ITEM_LINK_TO_CLIPBOARD_SUCCESS_MESSAGE =
   'Link is successfully copied!';
-export const COPY_ITEM_LINK_TO_CLIPBOARD_ERROR_MESSAGE =
-  'An error occurred while copying the link';
 export const RECYCLE_ITEMS_SUCCESS_MESSAGE =
   'The item(s) were successfully moved to the recycle bin.';
-export const RECYCLE_ITEMS_ERROR_MESSAGE =
-  'An error occurred while moving the item(s) to the recycle bin.';
 export const UPLOAD_ITEM_THUMBNAIL_SUCCESS_MESSAGE =
   'The thumbnail was successfully uploaded.';
 export const UPLOAD_ITEM_THUMBNAIL_FAILURE_MESSAGE =
